Revert "feature: check auth on sensors referenced in flex-context"
This reverts commit 8a3e1120b167d1ebf55bcd7ab2b13361baa8109a.
