+++ flexmeasures/conftest.py
@@ -9,7 +9,6 @@ from isodate import parse_duration
 import pandas as pd
 import numpy as np
 from flask import request, jsonify
-from flask_login import login_user, logout_user
 from flask_sqlalchemy import SQLAlchemy
 from flask_security import roles_accepted
 from pytest_mock import MockerFixture
@@ -23,7 +22,6 @@ from werkzeug.exceptions import (
     Gone,
 )
 
-from flexmeasures.api.tests.utils import UserContext
 from flexmeasures.app import create as create_app
 from flexmeasures.auth.policy import ADMIN_ROLE, ADMIN_READER_ROLE
 from flexmeasures.data.services.users import create_user
@@ -1364,37 +1362,3 @@ def add_beliefs(
 @pytest.fixture
 def mock_get_status(mocker: MockerFixture):
     return mocker.patch("flexmeasures.data.services.sensors.get_status", autospec=True)
-
-
-@pytest.fixture
-def requesting_user(request):
-    """Use this fixture to log in a user for the scope of a test.
-
-    Sets the user by passing it an email address (see usage examples below), or pass None to get the AnonymousUser.
-    Passes the user object to the test.
-    Logs the user out after the test ran.
-
-    Usage:
-
-    >>> @pytest.mark.parametrize("requesting_user", ["test_prosumer_user_2@seita.nl", None], indirect=True)
-
-    Or in combination with other parameters:
-
-    @pytest.mark.parametrize(
-        "requesting_user, status_code",
-        [
-            (None, 401),
-            ("test_prosumer_user_2@seita.nl", 200),
-        ],
-        indirect=["requesting_user"],
-    )
-
-    """
-    email = request.param
-    if email is not None:
-        with UserContext(email) as user:
-            login_user(user)
-            yield user
-            logout_user()
-    else:
-        yield
