+++ flexmeasures/data/schemas/scheduling/__init__.py
@@ -1,8 +1,5 @@
-from marshmallow import Schema, fields, validate, validates_schema, ValidationError
-from werkzeug.exceptions import Forbidden
+from marshmallow import Schema, fields, validate
 
-from flexmeasures.auth.policy import check_access
-from flexmeasures.data.schemas.scheduling.utils import find_sensors
 from flexmeasures.data.schemas.sensors import QuantityOrSensor, SensorIdField
 
 
@@ -32,15 +29,3 @@ class FlexContextSchema(Schema):
     inflexible_device_sensors = fields.List(
         SensorIdField(), data_key="inflexible-device-sensors"
     )
-
-    @validates_schema
-    def check_read_access_on_sensors(self, data: dict, **kwargs):
-        sensors = find_sensors(data)
-        for sensor, field_name in sensors:
-            try:
-                check_access(context=sensor, permission="read")
-            except Forbidden:
-                raise ValidationError(
-                    message=f"User has no read access to sensor {sensor.id}.",
-                    field_name=self.fields[field_name].data_key,
-                )
