+++ flexmeasures/data/schemas/tests/conftest.py
@@ -1,80 +1,25 @@
 import pytest
 from datetime import timedelta
 
-from flask_security import SQLAlchemySessionUserDatastore, hash_password
-
-from flexmeasures import Sensor, User, UserRole
-from flexmeasures.data.models.time_series import TimedBelief
-from flexmeasures.data.models.user import Account
+from flexmeasures.data.models.time_series import Sensor, TimedBelief
 from flexmeasures.data.models.generic_assets import GenericAsset, GenericAssetType
 
 
 @pytest.fixture(scope="module")
-def dummy_accounts(db, app):
-    dummy_account_1 = Account(name="dummy account 1")
-    db.session.add(dummy_account_1)
-
-    dummy_account_2 = Account(name="dummy account 2")
-    db.session.add(dummy_account_2)
-
-    # Assign account IDs
-    db.session.flush()
-
-    return {
-        dummy_account_1.name: dummy_account_1,
-        dummy_account_2.name: dummy_account_2,
-    }
-
-
-@pytest.fixture(scope="module")
-def dummy_user(db, app, dummy_accounts):
-    user_datastore = SQLAlchemySessionUserDatastore(db.session, User, UserRole)
-    user = user_datastore.create_user(
-        username="dummy user",
-        email="dummy_user@seita.nl",
-        password=hash_password("testtest"),
-        account_id=dummy_accounts["dummy account 1"].id,
-        active=True,
-    )
-    return user
-
-
-@pytest.fixture(scope="module")
-def dummy_assets(db, app, dummy_accounts):
+def dummy_asset(db, app):
     dummy_asset_type = GenericAssetType(name="DummyGenericAssetType")
     db.session.add(dummy_asset_type)
 
-    dummy_asset_1 = GenericAsset(
-        name="dummy asset 1",
-        generic_asset_type=dummy_asset_type,
-        owner=dummy_accounts["dummy account 1"],
+    _dummy_asset = GenericAsset(
+        name="DummyGenericAsset", generic_asset_type=dummy_asset_type
     )
-    db.session.add(dummy_asset_1)
+    db.session.add(_dummy_asset)
 
-    dummy_asset_2 = GenericAsset(
-        name="dummy asset 2",
-        generic_asset_type=dummy_asset_type,
-        owner=dummy_accounts["dummy account 1"],
-    )
-    db.session.add(dummy_asset_2)
-
-    dummy_asset_3 = GenericAsset(
-        name="dummy asset 3",
-        generic_asset_type=dummy_asset_type,
-        owner=dummy_accounts["dummy account 2"],
-    )
-    db.session.add(dummy_asset_3)
-
-    return {
-        dummy_asset_1.name: dummy_asset_1,
-        dummy_asset_2.name: dummy_asset_2,
-        dummy_asset_3.name: dummy_asset_3,
-    }
+    return _dummy_asset
 
 
 @pytest.fixture(scope="module")
-def setup_dummy_sensors(db, app, dummy_assets):
-    dummy_asset = dummy_assets["dummy asset 1"]
+def setup_dummy_sensors(db, app, dummy_asset):
     sensor1 = Sensor(
         "sensor 1",
         generic_asset=dummy_asset,
@@ -113,8 +58,7 @@ def setup_dummy_sensors(db, app, dummy_assets):
 
 
 @pytest.fixture(scope="module")
-def setup_efficiency_sensors(db, app, dummy_assets):
-    dummy_asset = dummy_assets["dummy asset 1"]
+def setup_efficiency_sensors(db, app, dummy_asset):
     sensor = Sensor(
         "efficiency",
         generic_asset=dummy_asset,
@@ -128,8 +72,7 @@ def setup_efficiency_sensors(db, app, dummy_assets):
 
 
 @pytest.fixture(scope="module")
-def setup_site_capacity_sensor(db, app, dummy_assets, setup_sources):
-    dummy_asset = dummy_assets["dummy asset 1"]
+def setup_site_capacity_sensor(db, app, dummy_asset, setup_sources):
     sensor = Sensor(
         "site-power-capacity",
         generic_asset=dummy_asset,
