+++ flexmeasures/data/schemas/tests/test_scheduling.py
@@ -239,10 +239,7 @@ def test_efficiency_pair(
         ),
     ],
 )
-@pytest.mark.parametrize(
-    "requesting_user", ["dummy_user@seita.nl"], indirect=True
-)
-def test_flex_context_schema(db, app, setup_site_capacity_sensor, flex_context, fails, dummy_user, requesting_user):
+def test_flex_context_schema(db, app, setup_site_capacity_sensor, flex_context, fails):
     schema = FlexContextSchema()
 
     # Replace sensor name with sensor ID
